Do not use Qt::SizeAllCursor
On Mac OS X, the cursor does not consist of four arrows pointing
in every direction, but is just a cross. Even worse, the cursor
does not have a proper anchor point, making the whole polygon
editing feature more or less unusable. The easiest solution seemed
to be to just not use it anymore.